// src/Utils/types.ts
export interface ImageType {
  src?: any;
  alt?: string;
  width?: number;
  height?: number;
  loading?: "lazy" | "eager";
}
// ...
export type TSubMenu = {
  id?: number;
  label?: string;
  path?: string;
  status?: string;
};

export type TMegaMenu = {
  id?: number;
  title?: string;
  submenu?: TSubMenu[];
  banner?: {
    path?: string;
    image?: ImageType;
  };
};